lots of reorganization for websafe build
