+++ cli/main.ts
@@ -1,12 +1,12 @@
 #!/usr/bin/env node
 
 import { Command } from "commander"
-import { fetchEasyEDAComponent } from "../lib/fetch-easyeda-json"
+import { fetchEasyEDAComponent } from "../lib/websafe/fetch-easyeda-json"
 import { convertEasyEdaJsonToTscircuitSoupJson } from "../lib/convert-easyeda-json-to-tscircuit-soup-json"
 import fs from "fs/promises"
 import packageJson from "../package.json"
 import { EasyEdaJsonSchema } from "lib/schemas/easy-eda-json-schema"
-import { convertRawEasyToTsx } from "lib/convert-to-typescript-component"
+import { convertRawEasyToTsx } from "lib/websafe/convert-to-typescript-component"
 import * as path from "path"
 import { normalizeManufacturerPartNumber } from "lib"
 import { convertEasyEdaJsonToVariousFormats } from "lib/convert-easyeda-json-to-various-formats"
