+++ lib/convert-easyeda-json-to-tscircuit-soup-json.ts
@@ -29,7 +29,7 @@ import { compose, scale, translate } from "transformation-matrix"
 import { computeCenterOffset } from "./compute-center-offset"
 import { mm } from "@tscircuit/mm"
 import { mil10ToMm } from "./utils/easyeda-unit-to-mm"
-import { normalizePinLabels } from "./normalize-pin-labels"
+import { normalizePinLabels } from "./websafe/normalize-pin-labels"
 
 const mil2mm = (mil: number | string) => {
   if (typeof mil === "number") return mm(`${mil}mil`)
