+++ lib/convert-easyeda-json-to-various-formats.ts
@@ -1,8 +1,8 @@
-import { fetchEasyEDAComponent } from "../lib/fetch-easyeda-json"
+import { fetchEasyEDAComponent } from "./websafe/fetch-easyeda-json"
 import { convertEasyEdaJsonToTscircuitSoupJson } from "../lib/convert-easyeda-json-to-tscircuit-soup-json"
 import fs from "fs/promises"
 import { EasyEdaJsonSchema } from "lib/schemas/easy-eda-json-schema"
-import { convertRawEasyToTsx } from "lib/convert-to-typescript-component"
+import { convertRawEasyToTsx } from "lib/websafe/convert-to-typescript-component"
 import * as path from "path"
 import { normalizeManufacturerPartNumber } from "lib"
