+++ lib/index.ts
@@ -1,6 +1,6 @@
 export * from "./convert-easyeda-json-to-tscircuit-soup-json"
-export * from "./fetch-easyeda-json"
-export { convertRawEasyToTsx as convertRawEasyEdaToTs } from "./convert-to-typescript-component"
+export * from "./websafe/fetch-easyeda-json"
+export { convertRawEasyToTsx as convertRawEasyEdaToTs } from "./websafe/convert-to-typescript-component"
 export { normalizeManufacturerPartNumber } from "./utils/normalize-manufacturer-part-number"
 export * from "./schemas/easy-eda-json-schema"
 export { convertEasyEdaJsonToVariousFormats } from "./convert-easyeda-json-to-various-formats"
