+++ lib/websafe/get-model-cdn-url.ts
