+++ lib/websafe/normalize-pin-labels.ts
