+++ tests/convert-to-ts/C14877.test.ts
@@ -1,6 +1,6 @@
 import { it, expect } from "bun:test"
 import atmegaRawEasy from "../assets/C14877.raweasy.json"
-import { convertBetterEasyToTsx } from "lib/convert-to-typescript-component"
+import { convertBetterEasyToTsx } from "lib/websafe/convert-to-typescript-component"
 import { EasyEdaJsonSchema } from "lib/schemas/easy-eda-json-schema"
 import { convertEasyEdaJsonToCircuitJson } from "lib"
