+++ tests/convert-to-ts/C165948-to-ts.test.ts
@@ -1,6 +1,6 @@
 import { it, expect } from "bun:test"
 import chipRawEasy from "../assets/C165948.raweasy.json"
-import { convertBetterEasyToTsx } from "lib/convert-to-typescript-component"
+import { convertBetterEasyToTsx } from "lib/websafe/convert-to-typescript-component"
 import { EasyEdaJsonSchema } from "lib/schemas/easy-eda-json-schema"
 import { convertEasyEdaJsonToCircuitJson } from "lib/convert-easyeda-json-to-tscircuit-soup-json"
 import { convertCircuitJsonToPcbSvg } from "circuit-to-svg"
