+++ tests/convert-to-ts/C2998002-to-ts.test.ts
@@ -1,6 +1,6 @@
 import { test, it, expect } from "bun:test"
 import chipRawEasy from "../assets/C2998002.raweasy.json"
-import { convertBetterEasyToTsx } from "lib/convert-to-typescript-component"
+import { convertBetterEasyToTsx } from "lib/websafe/convert-to-typescript-component"
 import { EasyEdaJsonSchema } from "lib/schemas/easy-eda-json-schema"
 import c2998002 from "tests/assets/C2998002.raweasy.json"
 import { convertCircuitJsonToPcbSvg } from "circuit-to-svg"
