+++ tests/convert-to-ts/C490691-to-ts.test.ts
@@ -1,6 +1,6 @@
 import { it, expect } from "bun:test"
 import chipRawEasy from "../assets/C490691.raweasy.json"
-import { convertBetterEasyToTsx } from "lib/convert-to-typescript-component"
+import { convertBetterEasyToTsx } from "lib/websafe/convert-to-typescript-component"
 import { EasyEdaJsonSchema } from "lib/schemas/easy-eda-json-schema"
 
 it("should convert C490691 into typescript file", async () => {
