+++ tests/fetch-tests/c2040.test.ts
@@ -1,5 +1,5 @@
 import { it, expect } from "bun:test"
-import { fetchEasyEDAComponent } from "lib/fetch-easyeda-json"
+import { fetchEasyEDAComponent } from "lib/websafe/fetch-easyeda-json"
 
 it("should fetch correct part for C2040", async () => {
   const result = await fetchEasyEDAComponent("C2040")
